Move UCIListener
bench 1666201
