+++ Serendipity/src/main/java/org/shawn/games/Serendipity/Search/Listeners/FinalReport.java
@@ -4,7 +4,7 @@ import com.github.bhlangonijr.chesslib.move.Move;
 
 public class FinalReport
 {
-	final Move bestMove;
+	public final Move bestMove;
 
 	public FinalReport(Move bestMove)
 	{
